Use EDB repository setup script on SUSE
Previously, we did not use the EDB repository setup script
on SUSE because it did not work on repeat deploys. Zypper would
raise because the repositories that the script attempts to install
already exist, and require unique names.

Now that the repository setup script task is skipped if the
repositories are already installed, this issue is not encountered.

Ensure repo tasks report changes only when actual change happened.

Fixes: TPA-974

Co-authored-by: Jonathan Renon <98327055+JonathanRenon-EDB@users.noreply.github.com>