Skip raft checks for BDR nodes with replica role
In this particular scenario, a BDR/PGD architecture incorporates
repmgr to handle physical replication of a subscriber-only BDR node
to another BDR node which acts as a backup. Due to previous
limitations, it was not possible to use two subscriber-only nodes.

The PGD upgrade process was failing because the physical replica BDR
node does not have RAFT enabled. As a fix, certain BDR-specific tasks
in the upgrade process now skip any node that has a replica role,
allowing for a successful upgrade.