This adds support for efm's jdbc.properties property
This was added to the efm template file in 5.0, even though it could
be specified earlier. Because it's not in the template, TPA can
set it and when the raw props are processed it will end up in the
final properties file. See TPA-921 for more information about why
this will only work with efm 5.0 and up.