Fix pgbouncer auth_query work for BDR4
Following the changes in TPA-382 as part of a388e43 commit, which
relocated `pgbouncer_get_auth()` to a dedicated
`pgbouncer_auth_database`, BDR4 deployments could encounter failures.
This occurred because the dedicated authentication database was not
consistently added to the `postgres_databases` list in the `init` role
for all BDR4 configurations. Previously, this dependency was masked as
the function resided in `pg_catalog` within every database.

This patch addresses this by ensuring the `pgbouncer_auth_database` is
reliably created for BDR4 deployments. Furthermore, creation of this
database is also ensured when HARP is configured, as `harp-proxy` may
utilize PgBouncer for connection management and thus also requires the
dedicated authentication database.

These adjustments guarantee that the `pgbouncer_auth_database`, and
consequently the `pgbouncer_get_auth()` function, are available as
needed for PgBouncer to function correctly in environments utilizing
BDR4 or HARP.

Fixes: TPA-382