Quote PEM_DB_PORT value when PEM is v10.1 or higher
PEM 10.1 adds support for multi-host connection strings from the web
application to the backend servers. To support this change, the
PEM_DB_PORT parameter in PEM's config_setup.py file is now a string
rather than an integer.

This commit moves the pem_schema_version check earlier so the result may
be used to determine whether to enclose the value in quotes (PEM 10.1+)
or not (earlier).

References: TPA-1009