Don't install pgbouncer db on pem servers
When we add the pgbouncer db to the list of databases to create, we
must exclude any db server that is a pem server. Fix the condition
that does this: the role for a pem server is 'pem-server' not
'pemserver'.

When we later check for the pgbouncer_get_auth function, similarly
we avoid checking on a pem server.

References: TPA-382