Use bdr_database for user creation when available
Use bdr_database as login_db parameter for the postgresql_user module
that is in charge of creating the postgres_users in your cluster.

This is needed by pgd especially in pgd6 to ensure that any change in
user configuration can be replicated (password change, grant, ...) in
the cluster.

We now verify if the bdr_database is in the actual database list before
deciding if we use it or default to postgres database. This allow M1 to
always use postgres and bdr cluster to use postgres on first deploy when
bdr_database is not yet created and switch to using bdr_database on
subsequent deploys.

References: TPA-951