Grant permissions to barman_role via privs module
Ensure the permissions for barman_role is idempotent and correctly uses
bdr_database when in a bdr scenario with bdr initialized.

Use postgresql_privs module to apply the grant on barman_role, so that
changes are only applied when needed.

Use the bdr_database on second deploys for BDR scenario. we want the DDL
to be replicated accross the cluster, the recommended way is to let bdr
do the replication and signal any change to the cluster as needed
(notify connection manager, reload configurations as needed, ...)

References: TPA-1020