Add task for add/drop role in tpaexec test
PGD replication of DDL needs to be checked to ensure a cluster is
working as expected after deployment.

Because ddl replication is on, once a role is created, this new one
should be made available for the other hosts right away. The same is
true when a role is dropped.

This change adds a set of task that adds a new role using the primary
bdr, checks that the role is correctly replicated accross the cluster
then removes it on the primary bdr node and verifies that it is no
longer available on the other nodes in the cluster.

References: TPA-1022