Ensure cluster name in integration tests is <63 characters
Save 6 characters on cluster names by prepending "c-" instead of
"cluster-" to the name that gets passed to the "Initialise Integration
Test" action; then truncate it to 63 characters just in case. This is
required because a cluster with a name 64 characters or more that uses
the BDR-Always-ON architecture fails during deployment when creating
BDR node groups (which are named after the cluster name).

Fixes: TPA-1027