This property will be added in efm 5.1, targeted for this August. The full property description hasn't been finalized yet, so it can't be included here. The property will be used to have efm attempt to rebuild a failed primary as a standby database, using pg_rewind when needed.
For testing: set the property to its non-default value in config.yml,
and without the change TPA doesn't write it out into the raw
properties, and the agent ends up with the default value.

Also tested the version check by setting the test to 5.2 (the dev
version of efm is 5.1) and the property value is ignored in that
case also.