Ensure pgpass_users correctly added to .pgpass file
Previously, any user in the postgres_users list specified with
generate_password: tru' AND included in the pgpass_users list
would NOT be added to the ~postgres/.pgpass file on the initial
deploy because the user's password did not yet exist when the
pgpass task was executed, thus the user was skipped.

This is fixed by invoking the pgpass task once more after all the
Postgres users have been created.

Furthermore, the repmgr user was previously part of the
default_pgpass_users list unless overridden by the pgpass_users list
in config.yml. Now it is only added to the  ~postgres/pgpass file as
part of the role/repmgr tasks, when repmgr is specified as the
failover_manager

The postgres_user (postgres or enterprisedb by default) is still part
of the default_pgpass_users list. If this is overridden by a
pgpass_users list in config.yml that does NOT include postgres_user,
a PEM-enabled cluster will fail to register agents as it needs the
encrypted postgres_user password.This is fixed by adding the
postgres_user to the ~postgres/.pgpass file as part of the PEM agent
tasks.

Fixes: TPA-158