Check cluster_vars dictionary during validate
Previously, if an invalid dictionary is set as the cluster_vars
dictionary in config.yml (such as cluster_vars variables referencing
other cluster_vars variables), TPA would swallow any Ansible errors
by falling back to the default value of an empty dictionary. This
resulted in every cluster_vars variable being undefined, so each was
set to it's TPA-default value. The resulting cluster would be
entirely different than what the user specified in their config.yml
file.

This is fixed by asserting that the cluster_vars dictionary is defined
and non-empty when the configuration file is loaded. Also as a final
bailout, the cluster_vars variable now no longer defaults to an empty
dictionary. This allows Ansible to throw an error when creating
group_vars and terminate.

Fixes: TPA-895, TPA-1033