Fix bare platform for pgd6 architectures
Configure command was failing for bare platform when using
PGD-X or PGD-S architecture, the platform bare was not correctly defined
in lib/tpa used by those architecures.

Adding a bare.yml in the platforms/ folder will ensure the platform is
recognized as a valid choice.

Add release note for TPA-1073

References: TPA-1073