Add generic upgrade command for PGD-X and PGD-S
Generate the main upgrade.yml file that is part of the commands
folder for the new PGD-S/X architectures. This is the main entry
point playbook that will evaluate the type of upgrade (major, minor),
create the required dynamic groups and include the appropriate
playbook for the cluster based on the upgrade path.

References: TPA-1037