Update dependencies for multiple archs
Besides x86_64 we're now dealing with multiple cpus. Adding a matrix
when updating the dependencies.

This matrix strategy is allowing to reutilise the same code we once had
for ppc64le when dealing with s390x cpu. Because we're using the same
edb-python version, we can use these requirements on el[89] and sles15.

References: TPA-1071