Skip deployment of a cluster on s390x
TPA can run on this arch, however, despite the code of different
EDB products (ie, efm or PGD) are on latest stages, the various
packages that the generated configuration needs are not available.

At the moment, we can just setup and provision. Once the packages are
available, we could revert this commit.