Dynamically generate Patroni ignore_slots for all Barman servers
Modify the Patroni configuration in TPA to dynamically generate the
ignore_slots list for physical replication slots based on hosts in the
role_barman group. This is especially important where more than one
backup servers (barman nodes) are part of the patroni cluster. This
makes sure patroni will not remove slots that any of the backup servers
need.

References: TPA-1087
Signed-off-by: Muhammad Haroon <muhammad.haroon@enterprisedb.com>