Don't enable routing for subscriber-only nodes
Fix a bug wereby subscriber-only groups would try to enable routing
during PGD6 deployments.

This property is not editable for such group since it is part of the
core definition of such group in PGD6.

The fix is to modify the conditional to avoid running this task when
facing a subscribers-only groups

References: TPA-1085