Add replication user to .pgpass on primary node
The replication_user entry is required the .pgpass file for both
replica and primary nodes managed by either efm or patroni as the
failover_manager. Previously, it was only added to the .pgass file
for replica nodes, resulting in connection issues to the primary
node after a switchover.

References: TPA-1100