Set pgd proxy and cli as bdr_package_version
This change copies `bdr_package_version` to either `pgd_proxy_package_version`
and/or `pgdcli_package_version` if any of them are not defined. Should a
customer defines any of them or both, that choice will be honoured
instead. That action allows current BDR customers to carry on with the
same cli and proxy versions everywhere unless it's decided to change.

references: TPA-938