Create postgres users defined in hostvars
In some situations, we cannot just take the `postgres_users` from one
host. `run_once` does that. When we have some corner cases where we
should have multiple users created in some hosts only, we need to
collect that information from ansible's hostvars instead.

The different filters allow to create a similar data structure to what
`postgres_users` provides so we can use it as a drop-in replacement.

Co-authored-by: Jonathan Renon <98327055+JonathanRenon-EDB@users.noreply.github.com>