Deploy steps on s390x architecture
EDB enabled native executors where this architecture can be tested. See
note below.

About repositories, the `enterprise` one  might be common for pgd-s and
pgd-x but we need different and extra repos for this architecture.

This commit tests the different packages BDR team has sent with the
various configurations we apply for amd64.

Note: We had to set max-jobs to 3 on s390x native executors because the
current infracture is brittle enough to overload it with sometimes just a
few jobs. We might need to tune this and even decrease the number if we get
messages like this in the logs:

FAILED! => {"changed": false, "cmd": "/usr/bin/systemctl",
"msg": "Failed to connect to bus: No such file or directory", "rc": 1,
"stderr": "Failed to connect to bus: No such file or directory\

In this context, the test containers are not able to run successfully.

references: TPA-1067
Co-authored-by: Jonathan Renon <98327055+JonathanRenon-EDB@users.noreply.github.com>