Add pgaudit package names for edbpge
Previously, adding pgaudit to the list of Postgres extensions with
the edbpge Postgres flavour would fail because the dictionary for
handling recognized extensions was missing an entry for edbpge and
it's corresponding package names. This entry is added, and additional
logic is incorporated to handle naming change for pgaudit packages
from version 16 onwards. We ensure that the correct package name is
being used for both open source users and edb repo users, for any
version of postgres.

References: TPA-997