Fix update_hosts into PDG-Always-ON minor upgrade
Ensure that minor upgrade of PGD5 can use update_hosts without issues by
porting the mechanism used in M1 upgrade into PGD5 minor upgrade.

We now generate a dynamic group g_update_hosts that is either the list
of host specified in update_hosts variable or the whole inventory if
the var is not set.

We then use this group to apply filtering on the various pieces of the
upgrade playbook as needed to ensure every task runs on the correct set
of nodes.

References: TPA-917