Add a default for `patroni_ssl_enabled` in the `init` role
This commit adds a default of `false` to `patroni_ssl_enabled` in
the `init` role.

That's required because we intend to use the value of
`patroni_ssl_enabled` when configuring the HAProxy nodes, and that
configuration step is performed before Patroni comes to the party.

The default is defined so the HAProxy configuration task won't fail
in cases where the user doesn't explicitly set `patroni_ssl_enabled`
in their `config.yml`.

References: TPA-1126.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>