Correctly configure HAProxy backends based on `patroni_ssl_enabled`
There was a bug in TPA when deploying HAProxy backends integrated
with Patroni clusters: if the Patroni cluster was configured to run
with SSL enabled (AKA `patroni_ssl_enabled: true`), HAProxy was still
being configured to perform HTTP health checks instead of HTTPS ones.

This commit fixes that issue. Now, when SSL is enabled in the Patroni
REST API server, HAProxy is configured with `check-ssl verify none`.
That effectively makes HAProxy perform HTTPS health checks.

Note: `verify none` is used because currently we don't set up a CA
certificate in the HAProxy nodes with capabilities of verifying the
server certificate of the Patroni nodes.

References: TPA-1126.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>