TPA-1163 Support for EFM release.vip.* properties
This adds support for three new EFM properties. The properties
control how the VIP is released in terms of timing and whether
or not it happens in the background while other things are
happening. Without these changes, there are cases where an
agent cannot make new database connections at the same time
the VIP is released.