Fix arch specific deps files update
- We rely on the workflow to update the .in files for s390x and ppc64 arch
we need to upload the .in files along with the .txt so that we have the
correct requirements in .in files for next runs of the workflow.
This ensure the PR will contain the updated .in files.
This also avoid future dependabot runs to revert changes in .txt files
by trying to match old requirements.

- There is a trailing white space in the version that fails the check
that compare target with actual requirements version.
this would always trigger the requirements to be rebuilt making the run
way longer than needed. we now add `xargs` to trim any such space.

- We now have both PPC and s390x specific dependencies.
Update the PR body to return a more accurate success message.

References: TPA-1165