Exclude BDR5 alter_node_option task for replicas
In a BDR5 cluster with a BDR node that is also a physical replica,
the upgrade process will fail during the alter_node_option task
because a BDR node is expected to have RAFT enabled, but the physical
replica BDR node does not. Skipping this task on any node with the
replica role avoids this issue and allows upgrade to complete
successfuly.

References: TPA-1158