Skip postgres/pgpass on PEM-agent Barman nodes
The postgres/pgpass task runs on pem-agent nodes in order to add the
postgres_user to the .pgpass file in this user's home directory.
However, this task fails on barman nodes that are also pem-agents,
since the directory does not exist.
This task is now skipped for nodes that do not have
postgres in their list of roles.

References: TPA-158, TPA-1171