Change configure output
Change the config.yml output for PGD-Always-ON cluster to be better
aligned with requirements for PGD5.9 and upgrade to 6+.

We need to avoid generating a config with both raft and
enable_proxy_routing  set to false since it will generate issues during
switch to connection manager on PGD 5.9.
This is due to the change of default in PGD5.9 for enable_proxy_routing
which is now set to True.
when TPA will run any deploy or upgrade with such config, it will fail
to alter the group option enable_proxy_routing, enable_raft being false
pgd won't let enable_proxy_routing be modified.

we now will generate a config.yml that sets:
for global routing:
- top group: raft enabled, routing enabled
- subgroups: raft enabled, routing disabled
- witness only group: raft disabled, routing unset

for local routing:
- top: raft enabled, routing disabled
- subgroups: raft enabled, routing enabled
- witness only group: raft disabled, routing unset

References: TPA-1168