Install bash-completion package from BDR5 onwards
This package allows pgd commands to auto complete on tab. Despite
that it's useful, there's at least one case where if it's not present
and the cluster includes Barman, during a `barman check` command, the
check might fail. This is because bash will request for the files
provided by the package and it won't fail silently making the ssh
output to include a single warning that Barman doesn't expect.

When that happen, you might see messages similar to:
$ barman check all
Server server_host:
	PostgreSQL: OK
	ssh: OK (PostgreSQL server)
        [...]
	ssh output clean: FAILED (the configured ssh_command must not add anything to the remote command output)

Instead of explicitly installing bash-completion, list it in packages
for pgd-cli and let the existing mechanism install it even on upgrade
scenario. This is compatible with the downloader.

references: TPA-1038