Setting backup_name misconfigures .pgpass file.
When setting the variable 'backup_name' in config.yml, the .pgpass file
located in the home directory for the 'barman' user, had the value of
the 'backup_name' variable instead of the instance's name it creates
the backup of (which is the default behaviour).

Now when the 'backup_name' variable is defined, it will change the name
of the barman 'conf' file, cron file, and directory, while preserving
the adequate behavior.

References: TPA-1052